avatar image scaled only if it is too big, and used in contact tooltips
Previously, if an avatar image was smaller than its frame, it was enlarged
to fill the frame, and that brought ugly results. Now, the avatar image is
not enlarged if it is smaller, but reduced if it is bigger than its frame;
it is fully visible, and not cut as previously.

Avatar images displayind in tooltips limited to 128*128px,
this fix first item of TEAM-460